+++ core/store/src/lib.rs
@@ -28,7 +28,6 @@ use crate::db::{
     refcount, DBIterator, DBOp, DBSlice, DBTransaction, Database, RocksDB, StoreStatistics,
     GENESIS_JSON_HASH_KEY, GENESIS_STATE_ROOTS_KEY,
 };
-pub use crate::trie::flat_state::FlatStateDelta;
 pub use crate::trie::iterator::TrieIterator;
 pub use crate::trie::update::{TrieUpdate, TrieUpdateIterator, TrieUpdateValuePtr};
 pub use crate::trie::{
@@ -36,6 +35,7 @@ pub use crate::trie::{
     Trie, TrieAccess, TrieCache, TrieCachingStorage, TrieChanges, TrieConfig, TrieStorage,
     WrappedTrieChanges,
 };
+pub use flat_state::FlatStateDelta;
 
 mod columns;
 pub mod config;
