+++ runtime/runtime/src/genesis.rs
@@ -14,10 +14,10 @@ use near_primitives::{
     trie_key::TrieKey,
     types::{AccountId, Balance, MerkleHash, ShardId, StateChangeCause, StateRoot},
 };
+use near_store::flat_state::FlatStateDelta;
 use near_store::{
     get_account, get_received_data, set, set_access_key, set_account, set_code,
-    set_postponed_receipt, set_received_data, FlatStateDelta, ShardTries, TrieUpdate,
-    WrappedTrieChanges,
+    set_postponed_receipt, set_received_data, ShardTries, TrieUpdate, WrappedTrieChanges,
 };
 
 use crate::config::RuntimeConfig;
